requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
Pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
matplotlib==3.8.0
numpy==1.24.3
torchtools==0.1.4
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.1
